make window focus a hook
