+++ packages/good-design/src/hooks/index.ts
@@ -1,2 +1,3 @@
 export * from "./useQueryParam";
 export { default as useScreenSize } from "./useScreenSize";
+export * from "./useWindowFocus";
