+++ packages/good-design/src/hooks/useWindowFocus.native.ts
@@ -1,0 +1,1 @@
+export {};
